Add snackbar to clip sharing
